SELC-4873 hide active product for operator in case he has no role on the product
